requirements.txt
flask
python-docx
openpyxl
qrcode[pil]
Pillow
docx2pdf
waitress
